+++ src/Backend/SgateShopgatePlugin/Plugin.php
@@ -5334,6 +5334,11 @@ class ShopgatePluginShopware extends ShopgatePlugin
                 ->setMaxResults($limit);
         }
 
+        // Export only items are in stock
+        if (!$this->config->getExportOutOfStockItems()) {
+            $builder->andWhere('(article.lastStock = 0 OR (mainDetail.inStock - mainDetail.stockMin > 0))');
+        }
+
         $builder->where('article.active = 1');
 
         // Currently we can only export stacking products which have minPurchase quantity to be a multiple of the purchaseSteps