+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - export right item number for order items
 - order debit payment method for SW >= 5.0
 
+## Added
+- possibility to ignore out of stock items in export
+
 ## [2.9.87] - 2019-05-21
 ### Fixed
 - export correct item number for order items